requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comtypes==1.4.12
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dateparser==1.2.2
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.7.0
docopt==0.6.2
ecdsa==0.19.1
edge-tts==7.2.3
email-validator==2.3.0
fastapi==0.116.1
fastapi-cli==0.0.10
fastapi-cloud-cli==0.1.5
fonttools==4.59.2
frozenlist==1.7.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
holidays
idna==3.10
Jinja2==3.1.6
Js2Py==0.74
kiwisolver==1.4.9
langdetect==1.0.9
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
numpy==2.3.2
orjson==3.11.3
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
pipwin==0.5.2
platformdirs==4.4.0
playsound==1.3.0
propcache==0.3.2
psutil==7.0.0
pyasn1==0.6.1
PyAudio==0.2.14
pycaw==20240210
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pygame==2.6.1
Pygments==2.19.2
pyinstaller==6.15.0
pyinstaller-hooks-contrib==2025.8
pyjsparser==2.7.1
pyparsing==3.2.3
pypiwin32==223
PyPrind==2.11.3
pySmartDL==1.3.4
pystray==0.19.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyttsx3==2.99
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.35.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
starlette==0.47.3
sympy==1.14.0
tabulate==0.9.0
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.17.3
yarl==1.20.1

# --- Streaming ASR (offline) ---
webrtcvad>=2.0.10
vosk>=0.3.45
